pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "PyCamp_Bot"
version = "3.0"
dependencies = [
  "munch==4.0.0",
  "python-telegram-bot==21.10",
  "peewee==3.17.9",
  "sentry-sdk==2.22.0",
]
requires-python = "==3.11.*"
authors = [
  {name = "Pyar", email = "pyar@pyar.com"},
]
description = "Bot de telegram para organizar pycamp"
readme = "README.md"

[project.optional-dependencies]
dev = [
  "flake8==7.1.2",
  "freezegun==1.5.1",
  "pytest==8.3.4",
]
doc = [
  "sphinx==7.3.7",
  "sphinx_rtd_theme==2.0.0",
]
